# api/openapi/api.yaml
openapi: 3.1.0
info:
  title: Lacpass App API
  description: This is the official API for Lacpass mobile app.
  version: "1.0"
  contact: {}
servers:
  - url: http://localhost:9081
    description: Local development server
paths:
  /ips:
    get:
      summary: Fetch IPS from national node.
      description: Fetch IPS from national node using session access token user identifier.
      tags:
        - IPS FHIR
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/NodeNameHeader'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IpsBundleResponse'
                example:
                  resourceType: "Bundle"
                  id: "16272"
                  meta:
                    versionId: "1"
                    lastUpdated: "2026-05-03T15:57:02.929+00:00"
                    source: "#vfkIjwwT0b5he4jI"
                    profile:
                      - "http://profiles.ihe.net/PHARM/MEOW/StructureDefinition/MedicationOverview"
                  identifier:
                    system: "urn:ietf:rfc:3986"
                    value: "urn:uuid:68c15694-52f1-4171-bc01-e63d666d2a71"
                  type: "document"
                  timestamp: "2025-06-17T18:36:18+00:00"
                  entry:
                    - fullUrl: "urn:uuid:68c15694-52f1-4171-bc01-e63d666d2a72"
                      resource:
                        resourceType: "Composition"
                        id: "composition-example"
                        meta:
                          profile:
                            - "http://profiles.ihe.net/PHARM/MEOW/StructureDefinition/MedicationOverviewComposition"
                        status: "final"
                        type:
                          coding:
                            - system: "http://loinc.org"
                              code: "56445-0"
                              display: "Medication summary"
                        subject:
                          reference: "urn:uuid:68c15694-52f1-4171-bc01-e63d666d2a73"
                        date: "2025-06-17T18:36:18+00:00"
                        author:
                          - reference: "urn:uuid:68c15694-52f1-4171-bc01-e63d666d2a73"
                        title: "Medication Overview"
                    - fullUrl: "Patient/16270"
                      resource:
                        resourceType: "Patient"
                        id: "patient-example"
                        identifier:
                          - system: "http://hospital.example.org"
                            value: "12345"
                        name:
                          - family: "Doe"
                            given:
                              - "John"
                        gender: "male"
                        birthDate: "1970-01-01"
                    - fullUrl: "urn:uuid:68c15694-52f1-4171-bc01-e63d666d2a74"
                      resource:
                        resourceType: "MedicationStatement"
                        id: "medication-statement-1"
                        status: "active"
                        medicationCodeableConcept:
                          coding:
                            - system: "http://snomed.info/sct"
                              code: "376834005"
                              display: "Atorvastatin 20 mg oral tablet"
                        subject:
                          reference: "urn:uuid:68c15694-52f1-4171-bc01-e63d666d2a73"
                        effectivePeriod:
                          start: "2025-01-01"
                          end: "2025-12-31"
                        dateAsserted: "2025-06-17T18:36:18+00:00"
                        reasonCode:
                          - text: "Hypercholesterolemia"
                        dosage:
                          - text: "20 mg daily"
                    - fullUrl: "urn:uuid:68c15694-52f1-4171-bc01-e63d666d2a75"
                      resource:
                        resourceType: "MedicationStatement"
                        id: "medication-statement-2"
                        status: "active"
                        medicationCodeableConcept:
                          coding:
                            - system: "http://snomed.info/sct"
                              code: "387013003"
                              display: "Aspirin 81 mg oral tablet"
                        subject:
                          reference: "urn:uuid:68c15694-52f1-4171-bc01-e63d666d2a73"
                        effectivePeriod:
                          start: "2025-01-01"
                        dateAsserted: "2025-06-17T18:36:18+00:00"
                        reasonCode:
                          - text: "Prophylaxis"
                        dosage:
                          - text: "81 mg daily"

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                BadRequest:
                  summary: "Bad request"
                  value:
                    - error: "bad_request"
                      error_description: "Bad request"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                ExpiredToken:
                  summary: "Expired token"
                  value:
                    - error: "missing_authorization_header"
                      error_description: "Missing Authorization header in request"
                KeySetNotAvailable:
                  summary: "Key set not available"
                  value:
                    - error: "key_set_not_available"
                      error_description: "Key set is not available, please try again later"
                BadFormattedAuthorizationHeader:
                  summary: "Bad formatted authorization header"
                  value:
                    - error: "bad_formatted_authorization_header"
                      error_description: "Missing bearer prefix in authorization header"
                InvalidToken:
                  summary: "Invalid token"
                  value:
                    - error: "invalid_token"
                      error_description: "Invalid token or signature"
                TokenUserIdNotFound:
                  summary: "Token user ID not found"
                  value:
                    - error: "token_user_id_not_found"
                      error_description: "Token does not contain user identifier"
                TokenRealmAccessNotFound:
                  summary: "Token realm access not found"
                  value:
                    - error: "token_realm_access_not_found"
                      error_description: "Token does not contain realm access information"
                TokenRolesNotFound:
                  summary: "Token roles not found"
                  value:
                    - error: "token_roles_not_found"
                      error_description: "Token does not contain roles information"
                TokenInvalidRole:
                  summary: "Token invalid role"
                  value:
                    - error: "token_invalid_role"
                      error_description: "Token contains an invalid role format"
                TokenUserUUIDNotFound:
                  summary: "Token user UUID not found"
                  value:
                    - error: "token_user_uuid_not_found"
                      error_description: "Token does not contain user UUID"
                UserIdentifierNotFound:
                  summary: "User identifier found"
                  value:
                    - error: "user_identifier_not_found"
                      error_description: "User identifier not found in request context"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                IPSNotfound:
                  summary: "IPS not found"
                  value:
                    - error: "not_found"
                      error_description: "No IPS found for the user"
  /ips/icvp:
    get:
      summary: Generate ICVP certificate from an IPS
      description: Generate ICVP vaccination certificate using the id of an IPS and optionally the id of an immunization.
      tags:
        - IPS FHIR
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/NodeNameHeader'
        - name: bundleId
          in: query
          required: true
          description: IPS bundle id
          schema:
            type: string
        - name: immunizationId
          in: query
          required: false
          description: Immunization id
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ICVPResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                BadRequest:
                  summary: "Bad request"
                  value:
                    - error: "bad_request"
                      error_description: "Bad request"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                ExpiredToken:
                  summary: "Expired token"
                  value:
                    - error: "missing_authorization_header"
                      error_description: "Missing Authorization header in request"
                KeySetNotAvailable:
                  summary: "Key set not available"
                  value:
                    - error: "key_set_not_available"
                      error_description: "Key set is not available, please try again later"
                BadFormattedAuthorizationHeader:
                  summary: "Bad formatted authorization header"
                  value:
                    - error: "bad_formatted_authorization_header"
                      error_description: "Missing bearer prefix in authorization header"
                InvalidToken:
                  summary: "Invalid token"
                  value:
                    - error: "invalid_token"
                      error_description: "Invalid token or signature"
                TokenUserIdNotFound:
                  summary: "Token user ID not found"
                  value:
                    - error: "token_user_id_not_found"
                      error_description: "Token does not contain user identifier"
                TokenRealmAccessNotFound:
                  summary: "Token realm access not found"
                  value:
                    - error: "token_realm_access_not_found"
                      error_description: "Token does not contain realm access information"
                TokenRolesNotFound:
                  summary: "Token roles not found"
                  value:
                    - error: "token_roles_not_found"
                      error_description: "Token does not contain roles information"
                TokenInvalidRole:
                  summary: "Token invalid role"
                  value:
                    - error: "token_invalid_role"
                      error_description: "Token contains an invalid role format"
                TokenUserUUIDNotFound:
                  summary: "Token user UUID not found"
                  value:
                    - error: "token_user_uuid_not_found"
                      error_description: "Token does not contain user UUID"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                ExpiredToken:
                  summary: "Not found"
                  value:
                    - error: "not_found"
                      error_description: "Not found"
  /ips/merge:
    post:
      summary: Merge two IPS bundles into a unified IPS.
      description: Merge two FHIR R4 IPS bundles into a single one, removing redundancy.
      tags:
        - IPS FHIR
      security:
        - ApiKeyAuth: []
      requestBody:
        description: IPS bundles to merge
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeIPSRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                MalformedCurrentIPS:
                  summary: "Bad request"
                  value:
                    - error: "bad_request"
                      error_description: "Malformed current IPS"
                MalformedNewIPS:
                  summary: "Bad request"
                  value:
                    - error: "bad_request"
                      error_description: "Malformed new IPS"
                NoComposition:
                  summary: "No composition"
                  value:
                    - error: "bad_request"
                      error_description: "Current IPS does not have its composition"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                ExpiredToken:
                  summary: "Expired token"
                  value:
                    - error: "missing_authorization_header"
                      error_description: "Missing Authorization header in request"
                KeySetNotAvailable:
                  summary: "Key set not available"
                  value:
                    - error: "key_set_not_available"
                      error_description: "Key set is not available, please try again later"
                BadFormattedAuthorizationHeader:
                  summary: "Bad formatted authorization header"
                  value:
                    - error: "bad_formatted_authorization_header"
                      error_description: "Missing bearer prefix in authorization header"
                InvalidToken:
                  summary: "Invalid token"
                  value:
                    - error: "invalid_token"
                      error_description: "Invalid token or signature"
                TokenUserIdNotFound:
                  summary: "Token user ID not found"
                  value:
                    - error: "token_user_id_not_found"
                      error_description: "Token does not contain user identifier"
                TokenRealmAccessNotFound:
                  summary: "Token realm access not found"
                  value:
                    - error: "token_realm_access_not_found"
                      error_description: "Token does not contain realm access information"
                TokenRolesNotFound:
                  summary: "Token roles not found"
                  value:
                    - error: "token_roles_not_found"
                      error_description: "Token does not contain roles information"
                TokenInvalidRole:
                  summary: "Token invalid role"
                  value:
                    - error: "token_invalid_role"
                      error_description: "Token contains an invalid role format"
                TokenUserUUIDNotFound:
                  summary: "Token user UUID not found"
                  value:
                    - error: "token_user_uuid_not_found"
                      error_description: "Token does not contain user UUID"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                ExpiredToken:
                  summary: "Not found"
                  value:
                    - error: "not_found"
                      error_description: "Not found"
  /qr:
    post:
      summary: Create QR data.
      description: Create QR data from VHL issuance.
      tags:
        - IPS FHIR
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/NodeNameHeader'
      requestBody:
        description: Data parameters
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VhlRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VhlResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                BadRequest:
                  summary: "Content cannot be empty"
                  value:
                    - error: "invalid_request"
                      error_description: "Content cannot be empty"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                ExpiredToken:
                  summary: "Expired token"
                  value:
                    - error: "missing_authorization_header"
                      error_description: "Missing Authorization header in request"
                KeySetNotAvailable:
                  summary: "Key set not available"
                  value:
                    - error: "key_set_not_available"
                      error_description: "Key set is not available, please try again later"
                BadFormattedAuthorizationHeader:
                  summary: "Bad formatted authorization header"
                  value:
                    - error: "bad_formatted_authorization_header"
                      error_description: "Missing bearer prefix in authorization header"
                InvalidToken:
                  summary: "Invalid token"
                  value:
                    - error: "invalid_token"
                      error_description: "Invalid token or signature"
                TokenUserIdNotFound:
                  summary: "Token user ID not found"
                  value:
                    - error: "token_user_id_not_found"
                      error_description: "Token does not contain user identifier"
                TokenRealmAccessNotFound:
                  summary: "Token realm access not found"
                  value:
                    - error: "token_realm_access_not_found"
                      error_description: "Token does not contain realm access information"
                TokenRolesNotFound:
                  summary: "Token roles not found"
                  value:
                    - error: "token_roles_not_found"
                      error_description: "Token does not contain roles information"
                TokenInvalidRole:
                  summary: "Token invalid role"
                  value:
                    - error: "token_invalid_role"
                      error_description: "Token contains an invalid role format"
                TokenUserUUIDNotFound:
                  summary: "Token user UUID not found"
                  value:
                    - error: "token_user_uuid_not_found"
                      error_description: "Token does not contain user UUID"
        "404":
          description: Not Found
  /qr/fetch:
    post:
      summary: Get IPS Bundle with valid VHL QR.
      description: Get IPS Bundle using a valid VHL QR.
      tags:
        - IPS FHIR
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/NodeNameHeader'
      requestBody:
        description: Data parameters
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VhlGetRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                ExpiredToken:
                  summary: "Expired token"
                  value:
                    - error: "missing_authorization_header"
                      error_description: "Missing Authorization header in request"
                KeySetNotAvailable:
                  summary: "Key set not available"
                  value:
                    - error: "key_set_not_available"
                      error_description: "Key set is not available, please try again later"
                BadFormattedAuthorizationHeader:
                  summary: "Bad formatted authorization header"
                  value:
                    - error: "bad_formatted_authorization_header"
                      error_description: "Missing bearer prefix in authorization header"
                InvalidToken:
                  summary: "Invalid token"
                  value:
                    - error: "invalid_token"
                      error_description: "Invalid token or signature"
                TokenUserIdNotFound:
                  summary: "Token user ID not found"
                  value:
                    - error: "token_user_id_not_found"
                      error_description: "Token does not contain user identifier"
                TokenRealmAccessNotFound:
                  summary: "Token realm access not found"
                  value:
                    - error: "token_realm_access_not_found"
                      error_description: "Token does not contain realm access information"
                TokenRolesNotFound:
                  summary: "Token roles not found"
                  value:
                    - error: "token_roles_not_found"
                      error_description: "Token does not contain roles information"
                TokenInvalidRole:
                  summary: "Token invalid role"
                  value:
                    - error: "token_invalid_role"
                      error_description: "Token contains an invalid role format"
                TokenUserUUIDNotFound:
                  summary: "Token user UUID not found"
                  value:
                    - error: "token_user_uuid_not_found"
                      error_description: "Token does not contain user UUID"
        "502":
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseList'
              examples:
                ValidationError:
                  summary: "Validation error"
                  value:
                    - error: "validation_error"
                      error_description: "Validation server did not return a valid access URL"
                UnsuccessfulValidation:
                  summary: "Unsuccessful validation"
                  value:
                    - error: "unsuccessful_validation"
                      error_description: "Validation step unsuccessful. Code: {0} Description: {1}"
                InvalidManifest:
                  summary: "Invalid manifest"
                  value:
                    - error: "invalid_manifest_url"
                      error_description: "Manifest server returned invalid bundle url."
...
  /nodes:
    get:
      summary: "List all available nodes"
      description: "Returns a list of node ids and display names. Use the id value with the Node-Name request header if multiple nodes are enabled."
      tags:
        - Discovery
      responses:
        "200":
          description: "A list of node ids and display names."
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - name
                  properties:
                    id:
                      type: string
                      description: "ASCII-safe node identifier used in the Node-Name request header."
                      example: node-1
                    name:
                      type: string
                      description: "Human-readable node name for display."
                      example: "Node 1"
              example:
                - id: node-1
                  name: "Node 1"
                - id: node-2
                  name: "Node 2"
        "404":
          description: "Multiple nodes are not enabled."
...
components:
  parameters:
    NodeNameHeader:
      name: Node-Name
      in: header
      required: false
      description: Optional ASCII-safe node id to route the request to a configured national node when USE_MULTIPLE_NODES is enabled. Omit this header, or provide an unknown node id, to use the default service configuration.
      schema:
        type: string
      example: node-1
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    ICVPResponse:
      type: object
      properties:
        data:
          type: string
          example: "base64-encoded-certificate-data"
        payload:
          type: object
          additionalProperties: true
          example:
            issuer: "Ministry of Health"
            valid_until: "2027-05-03"
...
    MergeIPSRequest:
      type: object
      properties:
        current_ips:
          type: object
          $ref: '#/components/schemas/IpsBundleResponse'
        new_ips:
          type: object
          $ref: '#/components/schemas/IpsBundleResponse'
...
    VhlGetRequest:
      type: object
      properties:
        data:
          type: string
          example: "shc:/567..."
        pass_code:
          type: string
          example: "123456"
    VhlRequest:
      type: object
      properties:
        content:
          type: string
          example: "FHIR-JSON-Content"
        expires_on:
          type: string
          example: "2026-12-31T23:59:59Z"
        pass_code:
          type: string
          example: "secure-pass-99"
    VhlResponse:
      type: object
      properties:
        data:
          type: string
          example: "https://vhl.link/abc"
        payload:
          type: object
          additionalProperties: true
          example:
            status: "active"
...
    ErrorResponseList:
      type: array
      items:
        type: object
        required:
          - error
          - error_description
        properties:
          error:
            type: string
          error_description:
            type: string